functionalities done

+++ src/components/BlogItem.js
@@ -14,7 +14,11 @@ function BlogItem({ blog }) {
   async function deleteBlog(id) {
     try {
       const blogsDoc = doc(db, "blogs", id);
-      await deleteDoc(blogsDoc);
+
+      const proceed = window.confirm("Are you sure you want to delete");
+      if (proceed) {
+        await deleteDoc(blogsDoc);
+      }
       navigate("/");
       // Display a success message, navigate back, or perform other actions.
     } catch (error) {
@@ -30,7 +34,7 @@ function BlogItem({ blog }) {
       <p>{blog.category}</p>
       <p>{blog.description}</p>
       <menu className={styles.actions}>
-        <Link>Edit</Link>
+        <Link to={`/${blog.id}/edit`}>Edit</Link>
         <button onClick={() => deleteBlog(blog.id)}>Delete</button>
       </menu>
     </article>
